fix(vite-ts): fixed the shadcn compatibility issues with vite when path-aliases were defined. Changes:
* Alias Issue during shadcn installation
current vite folder structure had 3 tsconfig.json files which confused shadcn. Shadcn expected 1 tsconfig and only looked for path aliases in the main tsconfig.json. When it didn't find the aliases, it added the folders with the name of aliases. To resolve this, path mappings were added to all 3 tsconfig.json files.

* Shad cn was added successfully

* npm-check-updates was used to updated all dependencies

* vite-tsconfig-paths package was installed incase the vite path resolution didn't work